Document persistence version constants across all storage layers
Added comprehensive documentation to all PERSISTENCE_FORMAT_VERSION and
liveStoreStorageFormatVersion constants explaining when to bump versions
to prevent accidental data loss.

**Changes:**
- Split system-tables.ts into separate eventlog-tables.ts and state-tables.ts
  for better organization and clearer distinction between safe and critical tables
- Added detailed documentation to liveStoreStorageFormatVersion explaining
  when version bumps are required vs not required
- Documented Cloudflare sync provider PERSISTENCE_FORMAT_VERSION with current
  schema and clear bump criteria
- Documented Electric sync provider PERSISTENCE_FORMAT_VERSION with PostgreSQL
  schema details
- Distinguished between STATE tables (safe to modify, auto-migrated) and
  EVENTLOG tables (require version bump, cause data loss if changed)

**Key Documentation:**
- EVENTLOG tables: Changes require version bump - permanent data loss otherwise
- STATE tables: Safe to modify - automatically rebuilt from eventlog
- All version constants now clearly state impact of version changes (soft reset)
