+++ packages/@livestore/common/src/index.ts
@@ -7,7 +7,7 @@ export * from './materializer-helper.ts'
 export * from './otel.ts'
 export * from './rematerialize-from-eventlog.ts'
 export * from './schema/state/sqlite/query-builder/mod.ts'
-export * from './schema/state/sqlite/system-tables.ts'
+export * from './schema/state/sqlite/system-tables/mod.ts'
 export * from './schema-management/migrations.ts'
 export * as SqliteDbHelper from './sqlite-db-helper.ts'
 export * from './sync/index.ts'
