+++ packages/@livestore/common/src/schema/mod.ts
@@ -6,6 +6,6 @@ export * from './schema.ts'
 export * as State from './state/mod.ts'
 export { SqliteAst, SqliteDsl } from './state/sqlite/db-schema/mod.ts'
 export * from './state/sqlite/schema-helpers.ts'
-export * as SystemTables from './state/sqlite/system-tables.ts'
+export * as SystemTables from './state/sqlite/system-tables/mod.ts'
 export type { UnknownEvents } from './unknown-events.ts'
 export { normalizeUnknownEventHandling, resolveEventDef } from './unknown-events.ts'
