+++ packages/@livestore/common/src/schema/schema.ts
@@ -4,7 +4,7 @@ import type { MigrationOptions } from '../adapter-types.ts'
 import type { EventDef, EventDefRecord, Materializer } from './EventDef.ts'
 import { tableIsClientDocumentTable } from './state/sqlite/client-document-def.ts'
 import type { SqliteDsl } from './state/sqlite/db-schema/mod.ts'
-import { stateSystemTables } from './state/sqlite/system-tables.ts'
+import { stateSystemTables } from './state/sqlite/system-tables/state-tables.ts'
 import type { TableDef } from './state/sqlite/table-def.ts'
 import type { UnknownEvents } from './unknown-events.ts'
 import { normalizeUnknownEventHandling } from './unknown-events.ts'
