+++ packages/@livestore/common/src/schema/state/sqlite/mod.ts
@@ -5,7 +5,7 @@ import type { Materializer } from '../../EventDef.ts'
 import type { InternalState } from '../../schema.ts'
 import { ClientDocumentTableDefSymbol, tableIsClientDocumentTable } from './client-document-def.ts'
 import { SqliteAst } from './db-schema/mod.ts'
-import { stateSystemTables } from './system-tables.ts'
+import { stateSystemTables } from './system-tables/state-tables.ts'
 import type { TableDef, TableDefBase } from './table-def.ts'
 
 export * from '../../EventDef.ts'
