+++ packages/@livestore/common/src/schema/state/sqlite/system-tables/mod.ts
@@ -1,0 +1,2 @@
+export * from './eventlog-tables.ts'
+export * from './state-tables.ts'
